+++ src/Storage/Database/Schema/Types/CarbonDateType.php
@@ -1,12 +1,19 @@
 <?php
 
-namespace Bolt\Storage\Mapping\Type;
+namespace Bolt\Storage\Database\Schema\Types;
 
 use Carbon\Carbon;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Types\DateTimeType;
+use Doctrine\DBAL\Types\DateType;
 
-class CarbonDateTimeType extends DateTimeType
+/**
+ * Doctrine DateType using Carbon.
+ *
+ * @internal
+ *
+ * @author Gawain Lynch <gawain.lynch@gmail.com>
+ */
+class CarbonDateType extends DateType
 {
     /**
      * {@inheritdoc}
